Add sematic types to non-terminal expressions (bug)
When determining the type for the following expression,

var z int32 = x + int32(y)

this bug occurred:

(o) 1. Recognizes `x` as `AddExpr` with type `int32` ($$ =
"int32").

(o) 2. Recognizes `y` as `AddExpr` with type `float32` and then
changes it to `MulExpr` with type `int32` after int32 type cast.

(x) 3. When matching with rule `AddExpr = AddExpr + MulExpr`, `x`
suddenly inherits the old type of `y` and becomes `AddExpr` with
rule `float32`, why???
(This is confirmed when I subbed `float32` with an arbitrary string
and observed that the sematic value ($$) of `x`'s `AddExpr` also
changed with it.)