Fix changing non-terminal expression sematic type
This bug occurred because when I wanted to copy the sematic value
of a sub grouping to the result non-terminal, I assigned the ptr
of the sub grouping to $$ instead of copying the value using str
copy. This was not a problem if the sub grouping was a string
literal, since its address was always the same. However, when the
sub grouping was a string variable, the sematic value of the
result non-terminal changed with the string variable's.

E.g.
-- OK --
$$ = "int32"

-- BUG --
TYPE = "int32"
$$ = TYPE
TYPE = "float32" (sematic value of $$ now = "float32"!)

-- FIX --
TYPE = "int32"
strcpy($$, TYPE)
TYPE = "float32" (sematic value of $$ unchanged)